Default to creating new places items globally
The Places panel in general is global, with its contents shared across
all apps that use it. However when you manually create a new entry
using right-click > Add Entry… the resulting dialog defaults to creating
an app-local item that won't be shown in any other apps or the file
dialogs.

Other methods of creating items such as drag-and-drop create items
globally; only this method does it locally.

This violates the otherwise-global nature of the Places panel, so it
isn't the best default. Let's change it.

BUG: 498579
FIXED-IN: 6.11