Make "Empty trash" icons red
This is a destructive action, and the HIG specifies that icons for
destructive actions should be colored red.

The current icons are black, inconsistent, and semantically incorrect.

edit-delete is red in the Breeze icon theme, and also more semantically
correct since the items in the trash will be deleted. Let's use this
icon instead.