Fix symlink path resolution in File Properties
If a symlink is opened from a virtual path like "desktop:/", the path
resolution incorrectly applied absolute UNIX path to the virtual path.

It resulted in an error whenever an "Open directory of symlink target"
button was pressed. The absolute symlink path was combined with a
protocol: "desktop:/absolute/symlink/path".

Always change the source URL to the Most Local URL first before resolving
the symlink.

BUG: 476808