Fail loading admin.so outside of the KIO PLUGINDIR
In 01fc16733f8a1ff0447c5aaa1e78f22faaceef1a I implemented a check
to make sure that the admin worker/plugin is not modifiable without
elevated privileges. This was to make sure that this worker, which
typically asks for elevated privileges, could not be utilized for
a privilege escalation attack.

During a security audit by Radically Open Security, Harry Pantazis
found a way to work around that security check by quickly replacing
a file between the security check and the use. A so-called "Time of
Check - Time of Use" vulnerability. He managed to use this
vulnerability with the right preparation 4 out of 141 attempts.

This commit aims to protect KIO from this attack pattern. This is
done by checking if the admin worker is installed in the standard
plugin location for KIO workers. This is considered sufficient
because a system-installed KIO library should not be modifiable
without elevated privileges and the same would normally also be
true for its default PLUGINDIR, where all the basic KIO workers
are loaded from. In a correctly configured system there should be
no way to circumvent this (and if there was, the KIO library would
already be insecure and anything we do here would already be for
naught).

In a development environment the PLUGINDIR is typically similarly
insecure as the KIO library itself. A side-effect of this commit is
that any system-installed admin worker will typically not be
loadable from the in-development KIO library anymore.

This commit also guards against a security concern pertaining to
the various ways library paths can be modified in Qt. See
QCoreApplication::libraryPaths() for details.