PasteDialog: make MIME type handling fancier
If the default MIME type is valid, pre-fill the text field with the
appropriate filename extension and move the cursor before it (allowing
the user to start typing or replace the suggestion immediately, without
impacting the extension).

When the user changes the MIME type using the combobox, change the
file extension in the text field if possible.

To facilitate this, pre-existing code to generate fancy combobox labels
was moved into the private component so that it could easily access the
raw strings used to generate them.

BUG: 499348
FIXED-IN: 6.13