KFileItem: fix stat() argument for URLs with local hostname
As reported in https://github.com/fish-shell/fish-shell/issues/11198,
reporting the working directory to Konsole via OSC 7
using a file:// URL with a non-empty hostname[^*], e.g.

	printf '\x1b]7;file://%s/tmp\a' $HOSTNAME

will cause the context menu's "Open Folder With" entry to be replaced with
"Open With".

This is because our KFileItem object wrongly reports that it's not a directory
("const bool isDir = m_props.isDirectory();"). 

This is due to the usage of QUrl::toLocalFile.
From https://doc.qt.io/qt-6/qurl.html#toLocalFile:

> If this URL contains a non-empty hostname, it will be encoded in
> the returned value in the form found on SMB networks (for example,
> "//servername/path/to/file.txt").

stat("//servername/path/to/file.txt") is invalid, which means we can't access
file metadata even if servername is the local hostname.

Fix that by stat'ing only QUrl::path(). This means that we also need
to make sure that we don't accidentally stat() a path from a URL like
"file://remote-server/bin".

[^*]: Note that it might seem odd to have a non-empty hostname
in a "file://" URL. But inside terminals, specifically for
OSC 7 and OSC 8 use cases, it seems to make sense, because the
application writing "file://" URLs can not generally assume that
the consumers of that URL runs on the same host (e.g. SSH). See also
https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda#file-uris-and-the-hostname