Add background to KUrlNavigator
This adds a background primitive to the KUrlNavigator, making it's default look similar as when edited.

This should make it easier for users to realise that the item is interactable.

This also changes the arrows between elements to clickable folder icons.

CCBUG:475106

See also https://invent.kde.org/system/dolphin/-/issues/68