KDirOperator: Make "newest first" the default kind of date sorting
When sorting by date, "newest first" is more useful in the context of a
file dialog where you're more likely to want to interact with newer
files rather than older ones.

To implement this, KDirOperator now inverts the sort order it's told to
apply when sorting by date. It has to be done this way because the
"reversed sorting" setting is stored globally, not per sort method.

FEATURE: 500512
FIXED-IN: 6.17