KNewFileMenu: Forbid saving file or folder with ~USERNAME as name
When creating new file or directory, check that any name which starts
with ~ does not resolve into a home directory on the machine.

This is due to an UNIX convention where "~user" can resolve into
"/home/user".

On Windows we do not have to worry about resolving this file.

Added an autotest for this as well.

CCBUG: 317513