KFileWidget and KDirOperator: Fix saving and loading preview settings
KDirOperator would load the iconSize after creating previewGenerator, which asks the iconSize from KDirOperator. This would make the iconSize always be invalid in previewGenerator.

KFileWidget would additionally toggle the action again, regardless of what the user setting is. It's better
just to explicitly follow what the user setting is and toggle the button accordingly. The previewGenerator value would always override what the user value is otherwise.

BUG: 501743