KProcessRunner: Fix launching actions without executable
If D-Bus activation is possible, it's not needed to have an executable,
since everything is done via D-Bus. Therefore look for the matching action
name instead of the exec value when launching an action. The specification
allows to omit the "Exec" key if "DBusActivatable" is true.

Most applications specify the Exec parameter for compatibility reasons, but
if the action's "Exec" line got removed from a D-Bus activatable desktop
file, then KIO just activates the application rather than calling the
requested action. This change fixes that issue.