KUrlNavigatorButton: Use arrow as separators
Due to the feedback of regular separator not being good enough for
separating folders, use an arrow separator instead.

This also refactors bunch of code around the button, so it should
be easier to tweak over time.

Remove icons due to the added clutter. They would work with full height chevrons but those are not technically feasible at this time.

BUG: 501587
BUG: 501575
BUG: 501582
BUG: 501589
BUG: 501706
BUG: 501803
FIXED-IN: 6.14