Show single-click selection emblem when using single-click mouse mode
In open/save dialog, we had no way to select items when using single-click mouse mode. (Except holding down control key).

This adds the same emblem that Dolphin uses for its selections.

It is shown in both treeviews and listviews, but only when single-click mode is active and the filewidget allows selecting multiple items.

BUG: 185793
FIXED-IN: 6.14