KFileWidget: Fix key navigation escaping in save dialogs
In save dialogs the keyboard navigation would escape during file
highlighting, since the fileHighlight sets the focus for fileName bar
for mouse operations.

This makes sure the user has to press Tab to explicitly escape the
keyboard navigation mode. For clicking the items, it should not affect
at all.

BUG: 466206
FIXED-IN: 6.14