core: Add infrastructure for inhibiting suspend in jobs
This calls the freedesktop Inhibit interface on DBus which will
inhibit suspend (but not display power management/screensaver).
When inside a sandbox it instead calls the XDG Desktop Portal
Inhibit interface.

When a job is destroyed or gets suspended, the inhibition is lifted.
When a job is resumed, `doInhibitSuspend` is called again to re-instate
the inhibition.

It is the job's responsibility to call `doInhibitSuspend` at
the appropriate time (e.g. in doStart/slotStart).