Redesign ExecutableFileOpenDialog
The "Open" vs "Execute" can be hard to grasp in certain languages.
Add the name and icon of the application that will open it to make
it clearer what will run it and what will just launch an app to
view it. Rename "Execute" to "Launch" according to HIG.

The prompt is adjusted to now ask for a choice when there is
only one (execute-only).

Also show the file icon or preview, name, and type to make a more
informed decision. Play the question message box sound, too.

While at it, turn it into a UI file.