KNewFileMenu: Offer custom folder icon directly from new dialog
This adds a strip of folder icons (row of colors and common custom
icons) to the new folder dialog to make it more convenient to
create folders with a custom icon and to raise awareness that this
is actually possible.

An icon that isn't in the default set can be picked through KIconDialog.
It is then added to the list, so it will gradually fill with the
user's most used custom icons. The "Custom Icon" collapsible group box
also remembers whether it was expanded, so users that use this
feature get to see it, whereas those who don't won't be bothered
by it.