Avoid unnecessary sequencing jobs in PreviewGenerator
Currently we keep constantly asking if the current item has sequences
support or not.

By sequences we mean things like hovering mouse over a folder thumbnail
and it goes through the files in it.

This MR will always run for the first sequence (the initial thumbnail)
but for the rest, then toggles a flag for the node with
Qt::DecorationPropertyRole. If the propertyrole is false, it skips any
further updates to avoid any unnecessary previewjob runs.