Add KIO::PreviewJob::useImage()
Currently, the PreviewJob provides previews as QPixmaps, however pixmaps
are not easy to integrate with QtQuick. One would have to convert the
QPixmap back to QImage.

This change introduces a new property that one could use to ask a PreviewJob
to provide previews as QImage objects to avoid intermediate QPixmaps.