Revert "KFileItemActions: check plugin instanciation"
This reverts commit 4c9b30955c3f684da3581b4a4a2322023254a102

As per discussion in KDE File Management matrix channel, we decided it's
better to revert this,
since this affects plugins that are on different release schedules: For
example if KIO updates before KIO-extras, this will break things.

What the reverted patch breaks:
- KIO Extras, FileItemLinkingPlugin: Destructor tries to use dangling
pointer, since the parent widget is already destroyed.
- Baloo Widgets, TagsFileItemAction: The plugin has a `m_menu` property
that gets destroyed due to it's parent getting destroyed, and it stays
around as dangling pointer.

If we just assign the contextmenu parent to these, they get destroyed
properly, though with the warnings the patch fixes.