Kfileitemactions: parent of plugins actions is the menu
they are added to, not the mainWindow.

This causes warnings such as:
```
qt.qpa.wayland: Creating a popup with a parent, QWidgetWindow(0x555dd14e8100, name="Dolphin#1Window") which does not match the current topmost grabbing popup, QWidgetWindow(0x555dd1cdafa0, name="DolphinContextMenuClassWindow") With some shell surface protocols, this is not allowed. The wayland QPA plugin is currently handling it by setting the parent to the topmost grabbing popup. Note, however, that this may cause positioning errors and popups closing unxpectedly. Please fix the transient parent of the popup.
```