PreviewJob: Kill the previewjob if it takes too long
In some cases thumbnailer may fail generating a thumbnail, be it
a bug in the plugin or the file is broken or various other reasons.

In those cases we keep trying to generate the thumbnail forever, without
timeout, leaving jobs running forever.

This patch adds a timer that kills the previewjob after 10 seconds.

CCBUG: 504067