DropJob: Emit popupMenuAboutToShow after setting things up
When adding actions to the popup, we were emitting popupMenuAboutToShow
too early, before we had the actions added. This could cause race
conditions, especially if there are plugin actions for the item.

We need to make sure our popup and its itemProps are done and ready
before emitting the signal. If we emit too early, we sometimes
ended up finishing making up the menu after DropJob::showMenu was
called, which would have nothing in it's m_menus set and thus do
nothing.

BUG: 506188