Support comparing checksums for multiple files
When the file properties dialog is opened with multiple files,
checksums are calculated and compared. Feedback is provided on whether
all checksums match, and the first file's checksum is shown.

The input field for custom checksums to verify against is disabled
when the dialog is opened for multiple files.

Results are cached.