KNewFileMenu: Add placeholder text
![image](/uploads/ed61f6b7dc9ea0fe4dcbb58449a3e8eb/image.png){width="440" height="141"}

It might be somewhat clear from context that the line edit is about entering the new file or folder name, but it isn't actually spelled out anywhere. E.g. for fully blind users the file or folder icon right next to the line edit is not noticable, so it might not be fully clear that a name for such a file type is entered there. Adding placeholder texts makes this perfectly clear though.

### Acknowledgement

This work is part of my project funded through the NGI0 Commons Fund, a fund established by NLnet with financial support from the European Commission's Next Generation Internet programme.