worker: add ability to inject mock workers
this allows us to create more bespoke workers that exhibit specific
behavior without having to rely on production workers.

for this purpose a new setTestWorkerFactory function was added that
takes a weak reference to a mock factory which may be used to create
mock workerbases. the bases will be used to power regular workerthreads,
same as would happen in production for the file worker.