Don't use FICLONE in special slow test path
We have a test that cancels a copy and checks that the partial files
is cleaned up. For that we have a special code path that introduces
artificial slowdown when a magically named file is copied, so the test
gets a chance to interrupt.

However when using the FICLONE feature available on some filesystems
the copy happens instantly and we have no chance to interrupt.

To address this skip the FICLONE codepath when in the special test mode