Avoid "Content-Length: 0" header for empty GET and HEAD requests
The QBuffer pointer eventually gets sent as a QIODevice ptr to
QNetworkAccessManager::sendCustomRequest(...) and Qt then sends a
"Content-Length: 0" header for all requests even if the data itself is
empty. Per RFC9110, "A user agent SHOULD NOT send a Content-Length
header field when the request message does not contain content and the
method semantics do not anticipate such data." Semantically, HTTP_GET
and HTTP_HEAD shouldn't send the header when the data is empty.
Account for the Qt behavior here until behavior is modified.

BUG: 507523