[kfilewidget] Fix saving file where same file without extension already exists
If we enter "foo" as filename, and have an automatic extension of .jpg, the
override check needs to check if foo.jpg exists, not foo.

However appendExtension() currently doesn't add the extension when the
file already exists, breaking the expectation here.

Remove that check to treat existing and non-existing basenames the same.

BUG: 377351