KFileFilterCombo: Fix initial delay in applying filter in KFileWidget
KFileWidget has a delay in applying the filter through editTextChanged, but filterChanged is instant. When setFilters is called on a KFileFilterCombo, we should emit filterChanged as we setCurrentIndex(), but we don't, causing it to go through the delayed editTextChanged path.

As a result of the issue, when a file dialog (e.g. 'Save As' in Gwenview) is opened, the results are shown unfiltered until 300ms elapses and the filter is applied.

This commit ensures the filter is applied immediately by emitting the right signal.

CCBUG: 508026