filepreviewjob: Load cached thumbnail in a worker thread
Loading even a small png still consumes the thread.

Given PreviewJob is an async API we can safely offload it to a worker
thread. We rely on PreviewJob to be limit concurrent activities.

If the FilePreviewJob is killed mid-operation, the worker thread will
continue loading the image but harmlessly discard the image at the end.