PreviewJob: Move plugin check to FilePreviewJob
Since FilePreviewJob already stats the item, we can get the mimetype
during that and avoid costly mimetype checking during PreviewJob
creation.

This change moves the mimetype and plugin checking from PreviewJob
to FilePreviewJob: First we stat the file, then use the MimeType
information to load the plugin we want to use for the preview
generation.