KUrlRequester: Announce preceding label when focused
Normally labels like "Working URL:" are announced previous to the
widget being labelled. However, previous to this commit this does
not work for KUrlRequester because the contained QLineEdit is only
a child of KUrlRequester. This commit makes sure the label which
KUrlRequester was setBuddy() for is announced when focus goes to
KUrlRequester. To accomplish this, the Label relation of the parent
KUrlRequester is mimiced by the internal QLineEdit by providing a
custom QAccessibleWidget for it.

This approach is preferable because it does not require any code
changes in user interfaces using this code.