KMountPoint: use mount id to match filesystem if possible
This should work with kernel 5.8+. This is the supposed way to match the
actual mount from /proc/self/mountinfo with the actual file. Not doing
it this way something like

747 253 0:79 / /media/smb/hostname/share rw,relatime shared:861 - autofs /etc/auto.master.d/auto.smb rw,fd=10,pgrp=1109,timeout=15,minproto=5,maxproto=5,offset,pipe_ino=11848
983 747 0:82 / /media/smb/hostname/share rw,nosuid,nodev,noexec,relatime shared:998 - cifs //hostname/share rw,vers=3.1.1,sec=krb5..

will show autofs, mounted from /etc.. in dolphin instead of the proper cifs and mounted from //hostname...

This was actually working some time ago, but commit id

35e575eb7d457c6e972ca373e2c408c31e4a7358

broke this since it uses the stat-ed device id for the fs entry insted the one from column 3
and thus it replaced the autofs dev id to 0:82 in the case above and made this the acutal mount insted of
the one below. This makes findByPath using the wrong actual "line".

There are reasons for 35e575eb7d457c6e972ca373e2c408c31e4a7358,
espescially btrfs and probaby bind mounts, but we might try as hard as
we want: the correct answer is mount id from statx, everything else is
guessing