Drop kfileaudiopreview
This is supposed to offer previews for audio files in the file dialog.

However this is shadowed by the KImageFilePreview, which is offering
previews for all files supported by previewjob, which includes audio files.

That means that for all intents and purposes this is dead code that never
gets used