filepreviewjob: Don't apply size check to directories
If a folder is situated on a bcachefs partition (or any other directory with 0 sized directory entry) its icon in Dolphin will not display thumbnails of its contents regardless of the preview options.

Judging by the log line `"PreviewJob: skipping an empty file, migth be a broken symlink"`, the size check should only apply to files not folders.

BUG: 507488 (Duplicate of BUG: 501882)