KFileWidget: Add filterWidget/locationEdit visibility, add allow pressing OK with empty location edit
Some applications, such as Ark, rely on allowing pressing OK
(or in Ark's case, Extract) even if the locationEdit is empty.
This is handled as current path in Ark.

This patch adds `setAllowOkWithEmptyLocationEdit` so that
the OK button will always be clickable, even with empty location.

It also adds two methods, that can be used to hide/show the
locationEdit and filterWidget widgets, and their labels.
In the case of Ark, neither are really needed.

CCBUG: 504421