[http] Remember that we're doing webdav while handling redirects
We were reverting to plain http(s), which confused clients like dolphin.
If the redirection target is plaintext use 'webdav' or 'dav' (respecting the current name choice), otherwise 'webdavs' or 'davs'.
Note that the encryption status of the redirection target can be different from the current one, so we cannot blindly use the old protocol.

BUG: 486790