trashimpl: better determine trashForMountPoint for network file systems
With uid mapping the local userid and remote userid might not match;
if permissions are checked by ownership, KIO won't consider a trash
directory even if it is writable.
This forced a fallback to the home trash which could mean moving a file
over a network, which is catastrophic for huge file sizes.

To fix this, instead of checking ownership and mode bits, use access()
which correctly queries the VFS for permission.

BUG:506755

Signed-off-by: Oliver Schramm <oliver.schramm97@gmail.com>