trash: don't fallback to home trash if trash is unavailable
If no trash can be created for the partition or network share, it might
not be desirable to move a file to the main trash on the home partition.
Especially when the partition is tmpfs (like /tmp), it would be
counterintuitive when the files would be persisted when it is "removed".
Even though this means, we can no longer test trash by using files
under /tmp.

So instead provide the user with a choice to delete the file
directly.

BUG:502315
BUG:481532
Signed-off-by: Oliver Schramm <oliver.schramm97@gmail.com>