KCoreDirLister: stop looking back at mount points
The removed code mentions an ancient bug from year 2002 and tries to work
around the deficiency of FAM. Back then watching a file using FAM was preventing
the mount point from being unmounted. More specifically, it was impossible to
unmount a USB flash drive while it was open in Konqueror.

Nowadays the established ways to monitor a file are inotify, kevent and stat.
Out of 3, both inotify and stat methods do not interfere with unmounting, so
we can safely keep watching the file regardles of its mount point. As for
kevent, this change doesn't really change anything - it is still impossible to
unmount a FS with kevent-watched file.