http: Use operation with known verb, if applicable
In Qt, HeadOperation (created using head()) does not behave
identical to a CustomOperation with verb "HEAD".

While a proper HeadOperation knows when not to expect any content,
a custom request with verb "HEAD" can get stuck waiting for content
indefinitely. This can for example happen when the server
requests authentication.

Custom operations are also never cached.

See https://codereview.qt-project.org/c/qt/qtbase/+/666393 but this
is only in dev, so it makes sense to address this downstream for now.

This work was sponsored by the government of Schleswig-Holstein,
under the assistence of Dataport.

BUG: 492178