Fix creating links to subdirectories of desktop links
This commit request fixes a bug that caused the creation of invalid links to
subdirectories of desktop links. The bug was caused by the use of a `desktop:/`
URL as the base for the new link, which resulted in an invalid path.

The fix is to use `mostLocalUrl()` to get the local file system path before
creating the link dialog. This ensures that the `KUrlRequester` start path
is valid and so removes the chance to create a non-working symlink.

BUG: 479350
FIXED-IN: 6.21