kfileplacesview: remove DropAction::MoveAction from the list of supportedActions
When we drag some item from "Places View" and drop it to some other destination
folder, there is `DropAction::MoveAction` allowed (on the drop popup menu) which
makes the source item is physically moved to destination folder and it is removed
from the source (but not from "Places View", so user even does not know the item
has been removed).


This commit corrects this behavior in two aspects:

1. It removes `DropAction::MoveAction` from the list of `supportedActions` in `KFilePlacesView::startDrag` method
2. It prevents putting plugin custom actions to drop popup menu as they can have
also Move-like destructive consequences

Now, when we drag some item from "Places View" to any other folder we just have
"Copy" and "Link" default actions in drop popup menu. Even when "Move by default"
setting is set, such drop popup menu is always displayed (and does not contain
Move-like actions).

BUG: 509231
FIXED-IN: 6.21