PreviewJob: Update previewJob data in determineNextFile, use given fileItems directly
This makes sure we always have the fresh data for the previewItem before
running the job.

We also make sure to insert the data we get from previewJob to the
maps, instead of overwriting them completely.

We also do not need to create another list of items, we can just use
KFileItemList directly and create a PreviewItem when it's needed.