Deprecate PreviewJob::removeItem(QUrl)
The last known user was found in K3FileIconView::removeItem() of kdelibs4.

Considering still any potential future users given the API already exist,
having an API which just allows to adjust the load of a running job by
removing items, but without the option to also add additional items or
querying the full active handled set and/or remaining items, as needed for
bookkeeping, seems incomplete for using such job(s) on the items of an
active view, where items are added & removed (if only to the visible part
of the scene).

So in the spirit "don't add API for future unknown needs", it seems
better to deprecate the current method to be removed on next occasion.
Once there are existing use-cases, the API can be (re-)added officially
as part of known needs.