PreviewJobTest: Add preview generation tests
This adds a simple PreviewJobTest that tests file generation.
For example, test if we can generate thumbnails from png files or not,
depending on what plugins we are using.

This also adds kio-extras to CI since kio does not have
any thumbnailers except the directory one, and we need more
for more robust testing.